requirements.txt
blinker==1.7.0
bpy==4.0.0
certifi==2024.2.2
charset-normalizer==3.1.0
click==8.1.7
cycler==0.11.0
Cython==3.0.10
debugpy==1.8.1
Flask==3.0.2
fonttools==4.38.0
fpdf==1.7.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.5.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.4
largestinteriorrectangle==0.2.0
llvmlite==0.39.1
MarkupSafe==2.1.5
mathutils==3.3.0
matplotlib==3.5.3
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
psutil==6.0.0
pyparsing==3.1.0
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
stitching==0.4.0
tifffile==2021.11.2
timm==0.5.4
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.3
Werkzeug==3.0.2
zipp==3.15.0
zstandard==0.22.0
